integrated P-smoother, still working on some issues

+++ inc/statistics.hpp
@@ -4,8 +4,8 @@
 
 std::vector<interval> merge_intervals(std::vector<interval> intervals);
 
-double compute_length_accuracy(std::string gt_fp, std::string rp_fp);
-double compute_accuracy(std::string gt_fp, std::string rp_fp);
-double compute_power(std::string gt_fp, std::string rp_fp);
-double compute_accumulative_power(std::string gt_fp, std::string rp_fp);
+float compute_length_accuracy(std::string gt_fp, std::string rp_fp);
+float compute_accuracy(std::string gt_fp, std::string rp_fp);
+float compute_power(std::string gt_fp, std::string rp_fp);
+float compute_accumulative_power(std::string gt_fp, std::string rp_fp);
